refactor: extract movies services to separated file
